Update the page nav function
