--- src/js/modules/util/pageNav.js
@@ -1,18 +1,0 @@
-const pageNav = (navItems, removeClass) => {
-  navItems.forEach((tab) => {
-    tab.addEventListener('click', (event) => {
-      if (event.target.textContent === 'Artworks') {
-        removeClass(navItems);
-        event.target.classList.add('active');
-      } else if (event.target.textContent === 'Exhibitions') {
-        removeClass(navItems);
-        event.target.classList.add('active');
-      } else if (event.target.textContent === 'Shop') {
-        removeClass(navItems);
-        event.target.classList.add('active');
-      }
-    });
-  });
-};
-
-export default pageNav;
